implemented LongList and LargePagedstore

+++ src/main/java/org/xenei/rdfstore/IdxData.java
@@ -1,10 +1,10 @@
 package org.xenei.rdfstore;
 
 public class IdxData<T> {
-    public final int idx;
+    public final long idx;
     public final T data;
     
-    public IdxData(int idx, T data) {
+    public IdxData(long idx, T data) {
         this.idx=idx;
         this.data=data;
     }
